Add @pure
This seems to be "hyperpure" since the final output is always the `===` same given the inputs. Without `Base.@pure`, it's not able to infer the full type because in the result of a static vector is its size. With `Base.@pure` it should be able to fully infer the type.